+++ src/SheetsServiceProvider.php
@@ -4,7 +4,8 @@ namespace Spatie\Sheets;
 
 use Illuminate\Support\ServiceProvider;
 use League\CommonMark\CommonMarkConverter;
-use Spatie\Sheets\ContentParsers\MarkdownParser;
+use League\CommonMark\ConverterInterface;
+use League\CommonMark\MarkdownConverterInterface;
 use Spatie\Sheets\ContentParsers\MarkdownWithFrontMatterParser;
 use Spatie\Sheets\PathParsers\SlugParser;
 use Spatie\Sheets\Repositories\FilesystemRepository;
@@ -30,17 +31,11 @@ class SheetsServiceProvider extends ServiceProvider
     {
         $this->mergeConfigFrom(__DIR__.'/../config/sheets.php', 'sheets');
 
-        $this->app->when(MarkdownWithFrontMatterParser::class)
-            ->needs(CommonMarkConverter::class)
-            ->give(function () {
-                return new CommonMarkConverter();
-            });
-
-        $this->app->when(MarkdownParser::class)
-            ->needs(CommonMarkConverter::class)
-            ->give(function () {
-                return new CommonMarkConverter();
-            });
+        if (! $this->app->bound(CommonMarkConverter::class)) {
+            $this->app->singleton(CommonMarkConverter::class);
+            $this->app->alias(CommonMarkConverter::class, MarkdownConverterInterface::class);
+            $this->app->alias(CommonMarkConverter::class, ConverterInterface::class);
+        }
 
         $this->app->singleton(Sheets::class, function () {
             $sheets = new Sheets();