+++ tests/TestCase.php
@@ -1,0 +1,16 @@
+<?php
+
+namespace Spatie\Sheets\Tests;
+
+use Orchestra\Testbench\TestCase as OrchestraTestCase;
+use Spatie\Sheets\SheetsServiceProvider;
+
+abstract class TestCase extends OrchestraTestCase
+{
+    public function getPackageProviders($app)
+    {
+        return [
+            SheetsServiceProvider::class,
+        ];
+    }
+}